+++ atprotolib-rs/src/types/app_bsky/actor/api_calls.rs
@@ -1,16 +1,16 @@
-use crate::{
-    api_calls::{AddApiAuth, ApiAuthConfig, ApiError},
-    types::app_bsky
+use super::{
+    api_responses::{SearchActorsResponse, SearchActorsTypeaheadResponse, SuggestionsResponse},
+    ProfileViewDetailed
 };
+use crate::api_calls::{AddApiAuth, ApiAuthConfig, ApiError};
 
-use super::{ProfileViewDetailed, SuggestionsResponse, SearchActorsTypeaheadResponse, SearchActorsResponse};
-
-/// Get a detailed profile view of an actor. Does not require auth, but contains relevant metadata with auth.
-/// 
+/// Get a detailed profile view of an actor. Does not require auth, but contains
+/// relevant metadata with auth.
+///
 /// <div class="warning">Requires the <code>apicalls</code> feature.</div>
-/// 
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to make the request to.
 /// * `api_auth_config` - The API authentication configuration.
 /// * `actor` - Handle or DID of the account to fetch the profile of.
@@ -24,7 +24,6 @@ pub async fn get_profile(
 
     let query_params = vec![("actor", actor)];
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -41,12 +40,13 @@ pub async fn get_profile(
     }
 }
 
-/// Get a list of suggested actors. Expected use is discovery of accounts to follow during new account onboarding.
-/// 
+/// Get a list of suggested actors. Expected use is discovery of accounts to
+/// follow during new account onboarding.
+///
 /// <div class="warning">Requires the <code>apicalls</code> feature.</div>
-/// 
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to make the request to.
 /// * `api_auth_config` - The API authentication configuration.
 /// * `limit` - The maximum number of actors to return. Defaults to 50.
@@ -69,7 +69,6 @@ pub async fn get_suggestions(
         query_params.push(("cursor", cursor));
     }
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -86,12 +85,13 @@ pub async fn get_suggestions(
     }
 }
 
-/// Find actor suggestions for a prefix search term. Expected use is for auto-completion during text field entry. Does not require auth.
-/// 
+/// Find actor suggestions for a prefix search term. Expected use is for
+/// auto-completion during text field entry. Does not require auth.
+///
 /// <div class="warning">Requires the <code>apicalls</code> feature.</div>
-/// 
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to make the request to.
 /// * `api_auth_config` - The API authentication configuration.
 /// * `query` - The search term to find actors for.
@@ -102,7 +102,7 @@ pub async fn search_actors_typeahead(
     api_auth_config: &ApiAuthConfig,
     query: &str,
     limit: Option<i32>
-) -> Result<app_bsky::actor::SearchActorsTypeaheadResponse, Box<dyn std::error::Error>> {
+) -> Result<SearchActorsTypeaheadResponse, Box<dyn std::error::Error>> {
     let api_url = format!(
         "https://{}/xrpc/app.bsky.actor.searchActorsTypeahead",
         host_name
@@ -114,7 +114,6 @@ pub async fn search_actors_typeahead(
     let limit_string = limit.unwrap_or_else(|| 10).to_string();
     query_params.push(("limit", limit_string.as_str()));
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -124,8 +123,7 @@ pub async fn search_actors_typeahead(
 
     match response.status() {
         reqwest::StatusCode::OK => {
-            let response_body: SearchActorsTypeaheadResponse =
-                response.json().await?;
+            let response_body: SearchActorsTypeaheadResponse = response.json().await?;
             Ok(response_body)
         }
         _ => Err(Box::new(ApiError::new(response).await?))
@@ -133,11 +131,11 @@ pub async fn search_actors_typeahead(
 }
 
 /// Find actors (profiles) matching search criteria. Does not require auth.
-/// 
+///
 /// <div class="warning">Requires the <code>apicalls</code> feature.</div>
-/// 
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to make the request to.
 /// * `api_auth_config` - The API authentication configuration.
 /// * `query` - The search term to find actors for.
@@ -163,7 +161,6 @@ pub async fn search_actors(
         query_params.push(("cursor", cursor));
     }
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
