+++ atprotolib-rs/src/types/app_bsky/actor/mod.rs
@@ -1,6 +1,8 @@
 #[cfg(feature = "apicalls")]
 pub mod api_calls;
 
+pub mod api_responses;
+
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
@@ -481,36 +483,4 @@ pub struct LabelerPrefItem {
     /// The DID of the labeler.
     #[serde(rename = "did")]
     pub did: String
-}
-
-/// A response to getting suggested actors.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct SuggestionsResponse {
-    /// A cursor for the stream.
-    #[serde(rename = "cursor", skip_serializing_if = "Option::is_none")]
-    pub cursor: Option<String>,
-
-    /// A list of actors.
-    #[serde(rename = "actors")]
-    pub actors: Vec<ProfileView>
-}
-
-/// A response for the typeahead search for actors.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct SearchActorsTypeaheadResponse {
-    /// A list of actors.
-    #[serde(rename = "actors")]
-    pub actors: Vec<ProfileViewBasic>
-}
-
-/// A response to searching for actors.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct SearchActorsResponse {
-    /// A cursor for the stream.
-    #[serde(rename = "cursor", skip_serializing_if = "Option::is_none")]
-    pub cursor: Option<String>,
-
-    /// A list of actors.
-    #[serde(rename = "actors")]
-    pub actors: Vec<ProfileView>
 }
